flask-app/requirements.txt
Flask==1.1.2
MarkupSafe==1.1.1
requests==2.24.0
Werkzeug==1.0.1
Pillow==7.0.0
numpy==1.19.2
